SKYW-2050 Update HTTP client
Updates the http client `get` method to accept a 'disposition'
parameter. This can be a value of 'json', 'text' or 'blob' and defaults
to 'json'. This will determine which fetch response method will be
called to resolve and cache the data returned from the call.
Dependencies also updated to latest versions.